Find the Week Day
Given the week day of nth day of a month, write a C program to find mth day of the month. For example, if it is given that 16th of the month is Wednesday then 25th of the month is Friday and if it is given that 29th of the month is Tuesday then 2nd of the month is Wednesday. Weekday number is given as input for ‘n’ that 0 indicates Sunday, 1 indicates Monday and so on.

Input Format

First line contains the value of n

Next line contains the week day number of ‘n’

Next line contains the value of ‘m’

Output Format

Print the week day of m. First letter of weekday should be in upper case